Make the session bar privacy pill the model switcher
The pill top right of the agent workspace ("Anonymous mode" / "Private
mode" / E2EE) now shows the current text model by name next to its
privacy mode, and clicking it opens the model picker in place instead
of the attestation page (still reachable from Settings and onboarding).
Selecting a model writes the global text-model setting and broadcasts
the existing provider-model-settings-changed event, so Settings and the
pill stay in sync from either direction.

ModelPickerDialog and its catalog helpers move verbatim from
AppSettings.tsx into their own module so the workspace can open the
picker without pulling in the settings surface.

Also syncs bun.lock with the border-beam/agentation dependencies that
landed in package.json with the onboarding redesign (#229).
